No longer assume a default format
When the user doesn't specify a --from format and auto-detection
fails, the official pandoc CLI assumes the format is Markdown.

That doesn't work so well for pandoc-tar, as in the future we may have
to support archives containing auxiliary files (images, stylesheets,
etc.) in addition to converted documents. The auto-detector would fail
to classify many/most/all auxiliary files, and it's not reasonable to
convert them as if they were Markdown documents.